api/requirements.txt
uvicorn==0.22.0
fastapi==0.96.0
python-dotenv==1.0.0
